+++ tiled/authn_database/migrations/versions/d829476bc173_split_apikeys_scope_into_create_and_.py
@@ -18,7 +18,7 @@ depends_on = None
 
 
 ROLES = ["admin", "user"]
-NEW_SCOPES = ["create:apikeys", "delete:apikeys", "create:node"]
+NEW_SCOPES = ["create:apikeys", "revoke:apikeys", "create:node"]
 OLD_SCOPES = ["apikeys", "create"]